fix: harden cmux sudo request flow
