+++ Sources/TerminalController.swift
@@ -1842,6 +1842,7 @@ class TerminalController {
         "system.top",
         "system.memory",
         "sudo.request",
+        "sudo.result",
     ]
 
     private nonisolated static func executionPolicy(forV2Method method: String) -> SocketCommandExecutionPolicy {
@@ -1955,6 +1956,8 @@ class TerminalController {
             return v2Result(id: request.id, v2SystemMemory(params: request.params))
         case "sudo.request":
             return v2Result(id: request.id, v2SudoRequestOnSocketWorker(params: request.params))
+        case "sudo.result":
+            return v2Result(id: request.id, v2SudoResultOnSocketWorker(params: request.params))
         case let method where method.hasPrefix("vm."):
             return socketWorkerCloudVMResponse(method: method, id: request.id, params: request.params)
         default:
