+++ docs/cmux-sudo.md
@@ -6,10 +6,12 @@
 
 1. The bundled CLI sends `sudo.request` over the cmux Unix socket with `argv`, `workspace_id`, `surface_id`, `caller_pid`, `caller_uid`, and `cwd`.
 2. The app validates the socket peer PID and UID, confirms the PID is a real cmux child process, reads the process environment, and rejects requests whose workspace or surface does not match the cmux terminal scope.
-3. The app shows the exact command to the user and calls `LAContext.evaluatePolicy(.deviceOwnerAuthentication, localizedReason:)`. Apple documents that this policy uses Touch ID and Apple Watch on macOS when available, then falls back to the user password.
-4. The app signs the helper payload with a per-app-session P-256 key and sends it to `/var/run/cmux-sudo-helper.sock`.
-5. The privileged helper validates the cmux app signature, verifies the signed payload, resolves the executable without a shell, runs one command, returns captured output and exit status, then exits.
-6. The app appends an audit entry whether the request is rejected, denied, or executed.
+3. The app records the request as pending and returns a request id immediately, so the socket worker is not blocked while the user is looking at a native authentication prompt.
+4. The CLI polls `sudo.result` with that request id until the app returns the terminal result. Result polling is accepted only from the same socket peer PID and UID that submitted the request.
+5. The app shows the exact command to the user and calls `LAContext.evaluatePolicy(.deviceOwnerAuthentication, localizedReason:)`. Apple documents that this policy uses Touch ID and Apple Watch on macOS when available, then falls back to the user password.
+6. The app signs the helper payload with a per-app-session P-256 key and sends it to `/var/run/cmux-sudo-helper.sock`.
+7. The privileged helper validates the cmux app signature and Team ID, verifies the signed payload, resolves the executable without a shell, runs one command, returns captured output and exit status, then exits.
+8. The app appends an audit entry whether the request is rejected, denied, or executed.
 
 ## Apple APIs
 
@@ -30,7 +32,7 @@ Production packaging still has to embed and sign the helper as an `SMAppService`
 
 - The LaunchDaemon must run as root and create `/var/run/cmux-sudo-helper.sock`.
 - The socket must be writable only by root and the local admin group (`0660`, `root:admin`).
-- The helper must validate the connecting process code signature before reading the request.
+- The helper must validate the connecting process code signature before reading the request. It accepts only cmux bundle identifiers signed by Manaflow Team ID `7WLXT3NR37`, including tagged debug, nightly, and staging bundle id suffixes.
 - The helper must accept only signed payloads from the cmux app process and must reject malformed payloads.
 - The helper must execute `argv` directly with `Process`, never through `/bin/sh` or a shell string.
 
@@ -59,6 +61,6 @@ An agent running inside cmux can ask for elevation, but it cannot grant elevatio
 
 A malicious or hallucinating agent can display a request for a dangerous command. The user still sees the exact command and must approve with device-owner authentication. Approval covers only that `argv`. The helper does not start an interactive root shell, does not cache approval, and does not allow shell escapes because it never executes a shell string.
 
-A compromised non-cmux process cannot use the app socket path to bypass validation because it will not be a cmux child with matching workspace and surface scope. A compromised same-user process that can inject code into the cmux app is out of scope; it can act as the app. Revocation is removing or unregistering the SMAppService daemon, deleting the helper socket, and restarting cmux.
+A compromised non-cmux process cannot use the app socket path to bypass validation because it will not be a cmux child with matching workspace and surface scope. It also cannot read another request's command output with only a guessed request id because result polling is pinned to the original socket peer. A compromised same-user process that can inject code into the cmux app is out of scope; it can act as the app. Revocation is removing or unregistering the SMAppService daemon, deleting the helper socket, and restarting cmux.
 
-The per-session signing key protects the payload between the app and helper from in-transit modification. It is not a substitute for helper-side code-signature validation because the public key travels with the payload.
+The per-session signing key protects the payload between the app and helper from in-transit modification. It is not a substitute for helper-side code-signature validation because the public key travels with the payload. The helper therefore also pins the connecting app to Manaflow's signing Team ID.
